addressing @delucchi-cmu comments on pr #82
