+++ src/hipscat_import/catalog/__init__.py
@@ -5,9 +5,6 @@ from .command_line_arguments import parse_command_line
 from .file_readers import (CsvReader, FitsReader, InputReader, ParquetReader,
                            get_file_reader)
 from .map_reduce import map_to_pixels, reduce_pixel_shards, split_pixels
-from .margin_cache import (generate_margin_cache,
-                           generate_margin_cache_with_client)
-from .margin_cache_arguments import MarginCacheArguments
 from .resume_files import (clean_resume_files, is_mapping_done,
                            is_reducing_done, read_histogram, read_mapping_keys,
                            read_reducing_keys, set_mapping_done,
